requirements.txt
python-dotenv
pymongo
pymongo[srv]==3.6
certifi
pandas
numpy
scikit-learn
##-e .